refactor: add GitHub Action
