+++ .github/workflows/build-and-test-stable.yml
@@ -1,0 +1,21 @@
+name: Build and Test Stable
+on:
+  workflow_dispatch:
+  push:
+    branches: [master]
+jobs:
+  build-and-test-stable:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout this repository
+        uses: actions/checkout@v2.3.4
+        with:
+          path: kiroku
+
+      - name: Add stable Debian repository and rosdep sources list
+        run: |
+          sudo apt update && sudo apt install curl
+          curl -s http://repository.ichiro-its.org/debian/setup.bash | bash -s
+          curl -s http://repository.ichiro-its.org/rosdep/setup.bash | bash -s
+      - name: Build and test workspace
+        uses: ichiro-its/ros2-build-and-test-action@main
