+++ .github/workflows/build-debian-nightly.yml
@@ -1,0 +1,46 @@
+name: Deploy Debian Nightly
+on:
+  workflow_dispatch:
+  push:
+    branches: [master]
+jobs:
+  deploy-debian-nightly:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout this repository
+        uses: actions/checkout@v2.3.4
+        with:
+          path: kiroku
+
+      - name: Add nightly Debian repository and rosdep sources list
+        run: |
+          sudo apt update && sudo apt install curl
+          curl -s http://repository.ichiro-its.org/debian/setup-nightly.bash | bash -s
+          curl -s http://repository.ichiro-its.org/rosdep/setup.bash | bash -s
+      - name: Build nightly Debian package
+        uses: ichiro-its/ros2-build-debian-action@main
+        with:
+          unique-version: true
+
+      - name: Deploy nightly Debian package to server
+        uses: appleboy/scp-action@master
+        with:
+          host: ${{ secrets.SSH_HOST }}
+          port: ${{ secrets.SSH_PORT }}
+          username: ${{ secrets.SSH_USER }}
+          password: ${{ secrets.SSH_PASS }}
+          source: "package/*.deb"
+          target: "~/temp/nightly/akushon/"
+          rm: true
+
+      - name: Prepare nightly Debian package in the server
+        uses: appleboy/ssh-action@master
+        with:
+          host: ${{ secrets.SSH_HOST }}
+          port: ${{ secrets.SSH_PORT }}
+          username: ${{ secrets.SSH_USER }}
+          password: ${{ secrets.SSH_PASS }}
+          script: |
+            cd ${{ secrets.SERVER_REPO_DIR }}/debian
+            reprepro includedeb nightly ~/temp/nightly/kiroku/package/*.deb
+            rm -rf ~/temp/nightly/kiroku/
