+++ .github/workflows/build-debian-stable.yml
@@ -1,0 +1,44 @@
+name: Deploy Debian Stable
+on:
+  workflow_dispatch:
+  release:
+    types: [created]
+jobs:
+  deploy-debian-stable:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout this repository
+        uses: actions/checkout@v2.3.4
+        with:
+          path: kiroku
+
+      - name: Add stable Debian repository and rosdep sources list
+        run: |
+          sudo apt update && sudo apt install curl
+          curl -s http://repository.ichiro-its.org/debian/setup.bash | bash -s
+          curl -s http://repository.ichiro-its.org/rosdep/setup.bash | bash -s
+      - name: Build stable Debian package
+        uses: ichiro-its/ros2-build-debian-action@main
+
+      - name: Deploy stable Debian package to server
+        uses: appleboy/scp-action@master
+        with:
+          host: ${{ secrets.SSH_HOST }}
+          port: ${{ secrets.SSH_PORT }}
+          username: ${{ secrets.SSH_USER }}
+          password: ${{ secrets.SSH_PASS }}
+          source: "package/*.deb"
+          target: "~/temp/stable/kiroku/"
+          rm: true
+
+      - name: Prepare stable Debian package in the server
+        uses: appleboy/ssh-action@master
+        with:
+          host: ${{ secrets.SSH_HOST }}
+          port: ${{ secrets.SSH_PORT }}
+          username: ${{ secrets.SSH_USER }}
+          password: ${{ secrets.SSH_PASS }}
+          script: |
+            cd ${{ secrets.SERVER_REPO_DIR }}/debian
+            reprepro includedeb stable ~/temp/stable/akushon/package/*.deb
+            rm -rf ~/temp/stable/kiroku/
