+++ package.xml
@@ -3,9 +3,9 @@
 <package format="3">
   <name>first_pubsub</name>
   <version>0.0.0</version>
-  <description>TODO: Package description</description>
+  <description>Logger Package</description>
   <maintainer email="azharanas173@gmail.com">azharanas17</maintainer>
-  <license>TODO: License declaration</license>
+  <license>MIT License</license>
 
   <depend>first_interfaces</depend>
   <depend>rclcpp</depend>
