+++ src/logger_pub_node.cpp
@@ -3,24 +3,24 @@
 
 #include "rclcpp/rclcpp.hpp"
 #include "first_interfaces/msg/log.hpp"
-#include "first_pubsub/logger_pub_node.hpp"
+#include "kiroku/logger_pub_node.hpp"
 
 using std::placeholders::_1;
 using namespace std::chrono_literals;
 
-  LoggerPub::LoggerPub()
-  : Node("logger_pub")
-  {
-    publisher_ = this->create_publisher<first_interfaces::msg::Log>("topic", 10);
-    timer_ = this->create_wall_timer(500ms, std::bind(&LoggerPub::topic_callback, this));
-  }
+LoggerPub::LoggerPub()
+: Node("logger_pub")
+{
+  publisher_ = this->create_publisher<first_interfaces::msg::Log>("topic", 10);
+  timer_ = this->create_wall_timer(500ms, std::bind(&LoggerPub::topic_callback, this));
+}
 
-  void LoggerPub::topic_callback() 
-  {
-    auto message = first_interfaces::msg::Log();
-    message.filename = "anaz";
-    message.message_log = "ppp";
-    message.level = "INFO";
-    message.time = "500";
-    publisher_->publish(message);
-  }
+void LoggerPub::topic_callback()
+{
+  auto message = first_interfaces::msg::Log();
+  message.filename = "logger";
+  message.message_log = "ini message";
+  message.level = "ERROR";
+  message.time = "500";
+  publisher_->publish(message);
+}
